feat(demo-bank): console seeding runs at boot, so a fresh store carries the full demo tenant
The rich automations/run-history/audit seeding lived only in the
seed:console CLI, which never runs inside a deployed container — a fresh
visitor on a local-store deploy saw two disabled automations and no
history. The seeding core moves to src/demo-script/console-seed.ts and
rides the same boot chain as seedDemoScript (and /api/demo/reset), still
idempotent; the CLI keeps working for pointing at other stores (--cloud,
--data-dir).
